Separate swing related class from view package.
