+++ src/main/java/oripa/persistence/doc/exporter/PictureExporter.java
@@ -35,8 +35,8 @@ import oripa.domain.creasepattern.CreasePattern;
 import oripa.geom.RectangleDomain;
 import oripa.gui.presenter.creasepattern.CreasePatternGraphicDrawer;
 import oripa.gui.view.creasepattern.ObjectGraphicDrawer;
-import oripa.gui.view.util.AffineCamera;
 import oripa.swing.drawer.java2d.CreasePatternObjectDrawer;
+import oripa.swing.view.util.AffineCamera;
 
 /**
  * @author Koji
