+++ src/main/java/oripa/swing/view/util/AffineCamera.java
@@ -16,7 +16,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package oripa.gui.view.util;
+package oripa.swing.view.util;
 
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseWheelEvent;
