+++ src/main/java/oripa/swing/view/util/GridBagConstraintsBuilder.java
@@ -1,4 +1,4 @@
-package oripa.gui.view.util;
+package oripa.swing.view.util;
 
 import java.awt.GridBagConstraints;
 import java.awt.Insets;
